ADD A USERBUILDER AND CREATE AUHTENTICATED USER #15
AuthenticatedUser will represent a connected user
